break out cycle resolution sync service tests
